Support for adb mouse with mousekeys
By keeping track of which source the mouse reports are coming from,
we're able to know if a mouse report without a button press is
because we released the button, or if the button is pressed on
another device.

The commit is rather dirty, with a magic number (source=1), and tons
of ifdef's.