Unicomp New Model M broken Pause/Break key fix
The PS/2 controller code for Unicomp's current generation PS/2 keyboard
sends the wrong scancodes when Ctrl + Pause (which should be Break) is
pressed.

Instead of it sending the Break sequence...

    E0 7E E0 F0 7E

... representing a press of Scroll Lock, which is Break because Ctrl is
being pressed, it instead sends...

    E0 77

... representing a press of Num Lock, which is just the old way of
getting Pause.

Since the E0 77 coming from the controller is not a normal sequence,
this workaround does not interfere with Ctrl + Num Lock, which still
Pauses as it should.