Refactor IBM PC lock state LED update routine
Keyboard lock state indicator LEDs are updated by sending a control
byte to the keyboard, which has one bit per individual LED state (the
rest of the byte being 0s). The order of the bits is different between
the USB HID and IBM PC/AT protocols, so they have to be reorganised
before TMK sends them to the keyboard.

The existing implementation of this works perfectly fine, but reorders
the LED state control byte to IBM format before sending it to a single
function that sends both the LED control enquiry byte (0xED) and the
actual LED state control byte itself, one after the other. However, not
every keyboard will respond to the enquiry byte with an acknowledgement
(0xFA), e.g. IBM terminal keyboards which will respond instead with an
0xFE meaning "RESEND", in which case the LED state control byte will
not be sent, meaning that it was just reorganised unnecessarily.

This refactored implementation splits the sending of the enquiry byte
into a separate function, so that the enquiry byte is sent first, and
the control byte is only reorganised if the keyboard has already
responded with the "ACK" acknowledgement byte.