Disable Mousekeys to prevent compiler crash
Even with this change, avr-gcc 8 and 9 are still unable to compile, but
it allows for a successful compilation in avr-gcc 5.4 and 5.5.

5.4 is still hanging on for dear life inside the VirtualBox VM provided
in the TMK wiki:
https://github.com/tmk/tmk_keyboard/wiki/Build-on-VirtualBox

5.5 is still currently available through Homebrew keg here:
https://github.com/osx-cross/homebrew-avr

I have not yet tested avr-gcc versions 6 or 7, because neither is
straightforwardly available to me (i.e. through on a package manager on
macOS, Windows, or Linux distro that can be set up quickly on a VM).