NEW FEATURE: ibmpc_usb converter lock state LEDs
Further restructured lock state LED routine to update both the keyboard
and any indicators wired to the converter itself. Tested on Unicomp New
Model M and various different IBM keyboards (a 5150 one, a 122-key, and
a 1990 Enhanced Keyboard). Current behaviour is that the converter LEDs
always light regardless of whether the keyboard has them, which is
primarily because of edge cases where a keyboard might respond
indicating that it can be sent LED state control bytes even though it
doesn't actually have LEDs with which to render the information.